ocr-service/requirements.txt
flask>=3.0,<4.0
paddlepaddle>=2.6,<3.0
paddleocr>=2.7,<3.0
pillow>=10.0,<11.0
gunicorn>=21.0,<23.0
